+++ lib/eventum/class.history.php
@@ -12,6 +12,8 @@
  */
 
 use Eventum\Db\DatabaseException;
+use Eventum\Event;
+use Eventum\EventDispatcher\EventManager;
 
 /**
  * Class to handle the business logic related to the history information for
@@ -69,10 +71,12 @@ class History
 
         $stmt = 'INSERT INTO {{%issue_history}} SET ' . DB_Helper::buildSet($params);
 
-        try {
-            DB_Helper::getInstance()->query($stmt, $params);
-        } catch (DatabaseException $e) {
-        }
+        DB_Helper::getInstance()->query($stmt, $params);
+        $params['his_id'] = DB_Helper::get_last_insert_id();
+        $params['prj_id'] = Auth::getCurrentProject();
+
+        $event = new Event\UnstructuredEvent($params);
+        EventManager::dispatch(Event\SystemEvents::HISTORY_ADD, $event);
     }
 
     /**